+++ src/Oro/Bundle/CheckoutBundle/Resources/public/js/app/components/transition-button-component.js
@@ -20,10 +20,12 @@ define(function(require) {
                 checkoutContent: '[data-role="checkout-content"]',
                 checkoutSummary: '[data-role="checkout-summary"]',
                 checkoutTotals: '[data-role="checkout-totals"]',
+                checkoutTitleRequiredLabel: '[data-role="checkout-title"] [data-role="require-label"]',
                 transitionTriggerContainer: '[data-role="transition-trigger-container"]',
                 transitionTrigger: '[data-role="transition-trigger"]',
                 stateToken: '[name$="[state_token]"]'
-            }
+            },
+            relatedCheckoutFormIds: []
         },
 
         /**
@@ -47,6 +49,10 @@ define(function(require) {
                 this.$form = this.getForm();
                 this.$form.bindFirst('submit.' + this.cid, this.preventSubmit.bind(this));
                 this.$form.on('submit.' + this.cid, this.onSubmit.bind(this));
+
+                this.getRelativeForms().forEach(
+                    form => [...form.find('[form]')].map(input => this.stashFormAttr(input))
+                );
             } else {
                 this.$el.on('click.' + this.cid, this.transit.bind(this));
             }
@@ -84,6 +90,8 @@ define(function(require) {
 
             this.$transitionTriggers.css('cursor', 'pointer');
             this.$transitionTriggers.on('click.' + this.cid, this.transit.bind(this));
+
+            this.$el.trigger('operation-button:init');
         },
 
         /**
@@ -111,9 +119,10 @@ define(function(require) {
                 mediator.execute('showFlashMessage', 'error', this.options.flashMessageOnSubmit);
                 return false;
             }
-            this.$form.validate();
 
-            if (this.$form.valid()) {
+            e.preventDefault();
+
+            if (this.formsValidate()) {
                 const eventData = {
                     stopped: false,
                     event: e
@@ -128,12 +137,50 @@ define(function(require) {
             }
         },
 
+        stashFormAttr(input) {
+            input.setAttribute('data-form', input.getAttribute('form'));
+            input.removeAttribute('form');
+
+            return input;
+        },
+
+        unStashFormAttr(input) {
+            input.setAttribute('form', input.getAttribute('data-form'));
+            input.removeAttribute('data-form');
+
+            return input;
+        },
+
+        formsValidate() {
+            return this.getAllForms().every(form => {
+                form.validate();
+                return form.valid();
+            });
+        },
+
+        getAllForms() {
+            return [this.$form, ...this.getRelativeForms()];
+        },
+
+        getRelativeForms() {
+            const forms = [];
+
+            if (this.options.relatedCheckoutFormIds) {
+                forms.push(...this.options.relatedCheckoutFormIds.map(
+                    relatedCheckoutFormId => $(`#${relatedCheckoutFormId}`)
+                ).filter(form => form.length));
+            }
+
+            return forms;
+        },
+
         /**
          * @param {Event} e
          * @param {Object} data
          */
         transit: function(e, data) {
             e.preventDefault();
+            e.stopPropagation();
             if (!this.options.enabled || this.inProgress || !this.options.transitionUrl) {
                 return;
             }
@@ -170,10 +217,14 @@ define(function(require) {
          * @returns FormData
          */
         getFormData: function() {
-            this.$form.find(this.options.selectors.stateToken)
-                .prop('disabled', false);
+            const formData = this.getAllForms().reduce((formData, form) => {
+                form.find(this.options.selectors.stateToken).prop('disabled', false);
+                for (const [name, value] of new FormData(form[0]).entries()) {
+                    formData.append(name, value);
+                }
+                return formData;
+            }, new FormData());
 
-            const formData = new FormData(this.$form[0]);
             if (!this.$form.find(':input').not(':button, :disabled').length) {
                 const formName = this.$form.attr('name') || 'oro_workflow_transition';
                 formData.append(`${formName}[]`, '');
@@ -207,8 +258,11 @@ define(function(require) {
                 const contentSelector = this.options.selectors.checkoutContent;
                 const summarySelector = this.options.selectors.checkoutSummary;
                 const totalsSelector = this.options.selectors.checkoutTotals;
+                const checkoutTitleRequiredLabelSelector = this.options.selectors.checkoutTitleRequiredLabel;
 
                 mediator.trigger('checkout-content:before-update');
+
+                $(checkoutTitleRequiredLabelSelector).remove();
 
                 const $sidebar = $(sidebarSelector);
                 $sidebar.html($response.find(sidebarSelector).html());
@@ -264,6 +318,12 @@ define(function(require) {
             if (this.$form) {
                 this.$form.off('.' + this.cid);
             }
+
+            this.getRelativeForms().forEach(
+                form => [...form.find('[form]')].map(input => this.unStashFormAttr(input))
+            );
+
+            this.$el.trigger('operation-button:dispose');
             this.$el.off('.' + this.cid);
             this.$transitionTriggers.off('.' + this.cid);