+++ src/Oro/Bundle/ShippingBundle/Resources/public/js/app/views/group-shipping-methods-view.js
@@ -6,6 +6,8 @@ define(function(require) {
     const $ = require('jquery');
     const NumberFormatter = require('orolocale/js/formatter/number');
     const mediator = require('oroui/js/mediator');
+    const multiShippingMethodsSelect2Template =
+        require('tpl-loader!oroshipping/templates/multi-shipping-methods-select2-template.html');
 
     const GroupShippingMethodsView = BaseView.extend({
         autoRender: true,
@@ -16,6 +18,10 @@ define(function(require) {
                 checkoutSummary: '[data-role="checkout-summary"]',
                 shippingMethodType: '[data-content="shipping_method_form"] [name^="shippingMethodType"]'
             }
+        },
+
+        events: {
+            'change select': 'onShippingMethodTypeChange'
         },
 
         /**
@@ -35,7 +41,6 @@ define(function(require) {
             this.options.template = _.template(this.options.template);
 
             mediator.on('transition:failed', this.render.bind(this, []));
-            mediator.on('checkout:shipping-method:rendered', this.onShippingMethodRendered, this);
         },
 
         /**
@@ -45,18 +50,29 @@ define(function(require) {
             this.updateShippingMethods(options);
             mediator.trigger('layout:adjustHeight');
             mediator.trigger('checkout:shipping-method:rendered');
-        },
 
-        onShippingMethodRendered: function() {
-            $(this.options.selectors.checkoutSummary).on(
-                'change',
-                this.options.selectors.shippingMethodType,
-                this.onShippingMethodTypeChange.bind(this)
-            );
+            const selectFormat = state => {
+                return multiShippingMethodsSelect2Template({
+                    ...state,
+                    ...$(state.element).data(),
+                    formatter: NumberFormatter
+                });
+            };
+
+            this.$('[data-role="select-shipping-method"]').inputWidget('create', 'select2', {
+                initializeOptions: {
+                    minimumResultsForSearch: -1,
+                    formatSelection: selectFormat,
+                    formatResult: selectFormat
+                }
+            });
+
+            return GroupShippingMethodsView.__super__.render.call(this);
         },
 
         onShippingMethodTypeChange: function(e) {
-            const methodType = $(e.target);
+            const data = $(e.target).inputWidget('data');
+            const methodType = $(data.element);
             const method = methodType.data('shipping-method');
             const type = methodType.data('shipping-type');
             const itemId = methodType.data('item-id');