+++ CHANGELOG.md
@@ -156,6 +156,9 @@ Previously known as GOD Object
 #### SaleBundle
 * Removed `oro_sale_accept_quote` as it is not used anywhere.
 
+#### CheckoutBundle
+* Removed `checkout_transition_submit` block type
+
 #### ProductBundle
 * Removed `oroproduct/js/app/views/sidebar-filters/filter-extra-hint-view`
 * Removed page_template values from `\Oro\Bundle\ProductBundle\Resources\views\layouts\default\config\page_templates.yml`.